Style modifications made and ternary relocated.
Per Nathan's comments, I have modified the appearance of the bars to more closely resemble the US version of the application.  Borders have been removed, elevation has been applied, colours have been tweaked to match the hex codes used in the US version of the application.  Additionally, the ternary statement which determined which colour the bar should display as has been relocated and is no longer determined in-line.